fix(download): Do not continue running target on incomplete download

+++ gonto/download.py
@@ -9,12 +9,12 @@ def http_get(
     allow_overwrite: bool = False,
     callback: Callable[[float], None] | None = None,
 ) -> None:
-    """Downdoad a file using HTTP protocol.
+    """Download a file using HTTP protocol.
 
     .. NOTE::
 
        The progress passed to callback function may stay to ``0`` during all
-       the download. This occures if the HTTP server does not provide the
+       the download. This occurs if the HTTP server does not provide the
        ``Content-Length``.
 
     :param url: The URL of the file to download.
@@ -30,6 +30,8 @@ def http_get(
 
     :raises FileExistsError: If the output file already exists and
         ``allow_overwrite`` is False.
+    :raise IOError: If the download was incomplete or if server sent more bytes
+        than expected.
     """
     output_file_path = Path(destination)
     output_file_path_partial = output_file_path.with_suffix(".part")
@@ -50,9 +52,26 @@ def http_get(
         while chunk := response.read(1024 * 1024):
             output_file.write(chunk)
             downloaded_length += len(chunk)
+            if total_length and downloaded_length > total_length:
+                raise IOError(
+                    "Downloaded more data than expected (%i/%i Bytes)"
+                    % (
+                        downloaded_length,
+                        total_length,
+                    )
+                )
             progress = downloaded_length / total_length if total_length else 0
             if callback:
                 callback(progress)
+
+    if total_length and downloaded_length != total_length:
+        raise IOError(
+            "Incomplete download (%i/%i Bytes)"
+            % (
+                downloaded_length,
+                total_length,
+            )
+        )
 
     if output_file_path.is_file():
         output_file_path.unlink()
